+++ corehq/warehouse/loaders/__init__.py
@@ -2,6 +2,7 @@ from corehq.warehouse.loaders.domain import DomainStagingLoader, DomainDimLoader
 from corehq.warehouse.loaders.form import FormStagingLoader, FormFactLoader
 from corehq.warehouse.loaders.group import GroupStagingLoader, GroupDimLoader
 from corehq.warehouse.loaders.location import LocationStagingLoader, LocationDimLoader
+from corehq.warehouse.loaders.synclog import SyncLogStagingLoader, SyncLogFactLoader
 from corehq.warehouse.loaders.user import UserStagingLoader, UserDimLoader
 
 
@@ -18,5 +19,7 @@ def get_loader_by_slug(slug):
 
         FormStagingLoader,
         FormFactLoader,
+        SyncLogStagingLoader,
+        SyncLogFactLoader,
     ]
     return {cls.slug: cls for cls in loaders}[slug]
